refactor: DefaultModal 컴포넌트에 Headless UI에서 제공하는 Dialog 적용
