+++ src/components/modal/DefaultModal.tsx
@@ -3,36 +3,41 @@ import { PropsWithChildren } from 'react';
 import XBigIcon from '@assets/svg/x_big.svg';
 import { styled } from 'stitches.config';
 import ModalBackground from './ModalBackground';
+import { Dialog } from '@headlessui/react';
 
 interface DefaultModalProps {
+  isModalOpened: boolean;
   title: string;
-  handleModalClose?: () => void;
+  handleModalClose: () => void;
 }
 
 const DefaultModal = ({
+  isModalOpened,
   title,
   handleModalClose,
   children,
 }: PropsWithChildren<DefaultModalProps>) => {
   return (
-    <>
+    <Dialog open={isModalOpened} onClose={handleModalClose}>
       <ModalBackground />
-      <SDefaultModal>
-        {title && (
+      <Dialog.Panel>
+        <SDialogWrapper>
           <SHeader>
-            <STitle>{title}</STitle>
-            <SXBigIcon onClick={handleModalClose} />
+            <Dialog.Title className="title">{title}</Dialog.Title>
+            <button onClick={handleModalClose}>
+              <SXBigIcon />
+            </button>
           </SHeader>
-        )}
-        <div>{children}</div>
-      </SDefaultModal>
-    </>
+          <div>{children}</div>
+        </SDialogWrapper>
+      </Dialog.Panel>
+    </Dialog>
   );
 };
 
 export default DefaultModal;
 
-const SDefaultModal = styled(Box, {
+const SDialogWrapper = styled(Box, {
   position: 'fixed',
   top: '50%',
   left: '50%',
@@ -50,12 +55,13 @@ const SHeader = styled(Box, {
   height: '$100',
   padding: '$40 $40 $36 $40',
   borderBottom: `1px solid $black40`,
-});
 
-const STitle = styled('p', {
-  width: '100%',
-  fontAg: '24_bold_100',
-  textAlign: 'center',
+  '.title': {
+    width: '100%',
+    fontAg: '24_bold_100',
+    textAlign: 'center',
+    color: '$white',
+  },
 });
 
 const SXBigIcon = styled(XBigIcon, {
